soc-ai-triage/requirements.txt
elasticsearch>=8.0.0,<9.0.0  # must be v8 — local ELK stack uses ES 8.x API headers
groq>=0.4.0
python-telegram-bot>=20.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
